Preserve original Message API
